+++ A-apps/the-boring-rpg/client-browser/CHANGELOG.md
@@ -11,8 +11,9 @@
 - 😷 chore: some internal refactors, bump dependencies
 
 
-## v0.
+## v0.63.1
 - 😍 feature: 16 new backgrounds (now totalling 114!
+- 🤩 feature: improved new version detection
 - 😅 fix: login and reload features are working again
 
 ## v0.62.1